Move some common helper into base class.
